Revert "README update & Code cleanup"
This reverts commit e90426324ce84316dc46f0ca5dc516716cb3331c.

+++ src/net/minecraft/server/EntityHuman.java
@@ -71,7 +71,7 @@ public abstract class EntityHuman extends EntityLiving {
 
     protected void b() {
         super.b();
-        this.datawatcher.a(16, (byte) 0);
+        this.datawatcher.a(16, Byte.valueOf((byte) 0));
     }
 
     public void m_() {
